feat: refactor usecase
