+++ app/src/main/java/com/thoughtworks/carapp/domain/hvac/AcOnStatusUseCase.kt
@@ -1,4 +1,4 @@
-package com.thoughtworks.carapp.domain.acmode
+package com.thoughtworks.carapp.domain.hvac
 
 import android.car.VehiclePropertyIds
 import android.car.hardware.property.CarPropertyManager
@@ -7,14 +7,22 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
-class GetAcOnStatusUseCase @Inject constructor(
+class AcOnStatusUseCase @Inject constructor(
     private val vehiclePropertyManager: VehiclePropertyManager
 ) {
-    operator fun invoke(): Flow<Boolean> {
+    fun getAcOnStatusFlow(): Flow<Boolean> {
         return vehiclePropertyManager.getPropertyFlow(
             VehiclePropertyIds.HVAC_AC_ON,
             CarPropertyManager.SENSOR_RATE_ONCHANGE
         )
             .map { it?.value as Boolean }
+    }
+
+    fun setAcOnStatus(value: Boolean) {
+        vehiclePropertyManager.setProperty(
+            VehiclePropertyIds.HVAC_AC_ON,
+            HvacAreas.ALL,
+            value
+        )
     }
 }
