+++ app/src/main/java/com/thoughtworks/carapp/presentation/carsetting/ac/AcViewModel.kt
@@ -1,15 +1,14 @@
 package com.thoughtworks.carapp.presentation.carsetting.ac
 
-import com.thoughtworks.carapp.domain.acmode.GetAcOnStatusUseCase
-import com.thoughtworks.carapp.domain.acmode.SetAcOnStatusUseCase
+import com.thoughtworks.carapp.domain.hvac.AcOnStatusUseCase
 import com.thoughtworks.carapp.domain.hvac.AcPowerStatusUseCase
 import com.thoughtworks.carapp.domain.hvac.GetAcAutoStatusUseCase
 import com.thoughtworks.carapp.domain.hvac.SetAcAutoStatusUseCase
 import com.thoughtworks.carapp.presentation.base.BaseViewModel
 import com.thoughtworks.carapp.presentation.base.Event
 import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
 import kotlinx.coroutines.flow.StateFlow
+import javax.inject.Inject
 
 sealed interface SettingScreenEvent : Event {
     object SwitchAcPowerEvent : SettingScreenEvent
@@ -20,14 +19,13 @@ sealed interface SettingScreenEvent : Event {
 @HiltViewModel
 class AcViewModel @Inject constructor(
     private val acPowerStatus: AcPowerStatusUseCase,
-    getAcAutoStatus: GetAcAutoStatusUseCase,
-    getAcOnStatus: GetAcOnStatusUseCase,
+    private val acOnStatus: AcOnStatusUseCase,
     private val setAcAutoStatus: SetAcAutoStatusUseCase,
-    private val setAcOnStatus: SetAcOnStatusUseCase,
-    ) : BaseViewModel<SettingScreenEvent>() {
+    getAcAutoStatus: GetAcAutoStatusUseCase,
+) : BaseViewModel<SettingScreenEvent>() {
     val isAcPowerOn: StateFlow<Boolean> = acPowerStatus.getAcPowerStatusFlow().stateWith(false)
     val isAcAutoOn: StateFlow<Boolean> = getAcAutoStatus().stateWith(false)
-    val isAcOn: StateFlow<Boolean> = getAcOnStatus().stateWith(false)
+    val isAcOn: StateFlow<Boolean> = acOnStatus.getAcOnStatusFlow().stateWith(false)
 
     override fun handleEvents(event: SettingScreenEvent) {
         when (event) {
@@ -39,7 +37,7 @@ class AcViewModel @Inject constructor(
             }
             SettingScreenEvent.SwitchAcOnEvent -> {
                 if (isAcPowerOn.value) {
-                    setAcOnStatus(isAcOn.value.not())
+                    acOnStatus.setAcOnStatus(isAcOn.value.not())
                 }
             }
         }
